requirements.txt
confluent_kafka
psycopg2-binary
